Rename meilisearch client package.
