+++ src/main/java/io/vanslog/spring/data/meilisearch/client/msc/package-info.java
@@ -20,4 +20,4 @@
  */
 @org.springframework.lang.NonNullApi
 @org.springframework.lang.NonNullFields
-package io.vanslog.spring.data.meilisearch.client.mlc;
+package io.vanslog.spring.data.meilisearch.client.msc;
