+++ src/main/java/io/vanslog/spring/data/meilisearch/core/MeilisearchTemplate.java
@@ -24,8 +24,8 @@ import io.vanslog.spring.data.meilisearch.TaskStatusException;
 import io.vanslog.spring.data.meilisearch.UncategorizedMeilisearchException;
 import io.vanslog.spring.data.meilisearch.annotations.Document;
 import io.vanslog.spring.data.meilisearch.client.MeilisearchClient;
-import io.vanslog.spring.data.meilisearch.client.mlc.RequestConverter;
-import io.vanslog.spring.data.meilisearch.client.mlc.ResponseConverter;
+import io.vanslog.spring.data.meilisearch.client.msc.RequestConverter;
+import io.vanslog.spring.data.meilisearch.client.msc.ResponseConverter;
 import io.vanslog.spring.data.meilisearch.core.convert.MappingMeilisearchConverter;
 import io.vanslog.spring.data.meilisearch.core.convert.MeilisearchConverter;
 import io.vanslog.spring.data.meilisearch.core.mapping.MeilisearchPersistentEntity;
